+++ client/extensions/woocommerce/state/action-types.js
@@ -13,9 +13,9 @@ export const WOOCOMMERCE_ERROR_SET = 'WOOCOMMERCE_ERROR_SET';
 export const WOOCOMMERCE_ERROR_CLEAR = 'WOOCOMMERCE_ERROR_CLEAR';
 export const WOOCOMMERCE_LOCATIONS_REQUEST = 'WOOCOMMERCE_LOCATIONS_REQUEST';
 export const WOOCOMMERCE_LOCATIONS_REQUEST_SUCCESS = 'WOOCOMMERCE_LOCATIONS_REQUEST_SUCCESS';
-export const WOOCOMMERCE_ORDER_REFUND_REQUEST = 'WOOCOMMERCE_ORDER_REFUND_REQUEST';
-export const WOOCOMMERCE_ORDER_REFUND_REQUEST_SUCCESS = 'WOOCOMMERCE_ORDER_REFUND_REQUEST_SUCCESS';
-export const WOOCOMMERCE_ORDER_REFUND_REQUEST_FAILURE = 'WOOCOMMERCE_ORDER_REFUND_REQUEST_FAILURE';
+export const WOOCOMMERCE_ORDER_REFUND_CREATE = 'WOOCOMMERCE_ORDER_REFUND_CREATE';
+export const WOOCOMMERCE_ORDER_REFUND_CREATE_SUCCESS = 'WOOCOMMERCE_ORDER_REFUND_CREATE_SUCCESS';
+export const WOOCOMMERCE_ORDER_REFUND_CREATE_FAILURE = 'WOOCOMMERCE_ORDER_REFUND_CREATE_FAILURE';
 export const WOOCOMMERCE_ORDER_REQUEST = 'WOOCOMMERCE_ORDER_REQUEST';
 export const WOOCOMMERCE_ORDER_REQUEST_FAILURE = 'WOOCOMMERCE_ORDER_REQUEST_FAILURE';
 export const WOOCOMMERCE_ORDER_REQUEST_SUCCESS = 'WOOCOMMERCE_ORDER_REQUEST_SUCCESS';
