+++ client/extensions/woocommerce/state/sites/orders/reducer.js
@@ -8,9 +8,6 @@ import { keyBy } from 'lodash';
  */
 import { combineReducers } from 'state/utils';
 import {
-	WOOCOMMERCE_ORDER_REFUND_REQUEST,
-	WOOCOMMERCE_ORDER_REFUND_REQUEST_FAILURE,
-	WOOCOMMERCE_ORDER_REFUND_REQUEST_SUCCESS,
 	WOOCOMMERCE_ORDER_REQUEST,
 	WOOCOMMERCE_ORDER_REQUEST_FAILURE,
 	WOOCOMMERCE_ORDER_REQUEST_SUCCESS,
@@ -18,6 +15,7 @@ import {
 	WOOCOMMERCE_ORDERS_REQUEST_FAILURE,
 	WOOCOMMERCE_ORDERS_REQUEST_SUCCESS,
 } from 'woocommerce/state/action-types';
+import refunds from './refunds/reducer';
 
 /**
  * Returns the updated order requests state after an action has been
@@ -54,26 +52,6 @@ export function isQueryLoading( state = {}, action ) {
 		case WOOCOMMERCE_ORDERS_REQUEST_SUCCESS:
 		case WOOCOMMERCE_ORDERS_REQUEST_FAILURE:
 			return Object.assign( {}, state, { [ `{page:${ action.page }}` ]: WOOCOMMERCE_ORDERS_REQUEST === action.type } );
-		default:
-			return state;
-	}
-}
-
-/**
- * Returns the updated order requests state after an action has been
- * dispatched. The state reflects a mapping of query (page number) to a
- * boolean reflecting whether a request for that page is in progress.
- *
- * @param  {Object} state  Current state
- * @param  {Object} action Action payload
- * @return {Object}        Updated state
- */
-export function isRefunding( state = {}, action ) {
-	switch ( action.type ) {
-		case WOOCOMMERCE_ORDER_REFUND_REQUEST:
-		case WOOCOMMERCE_ORDER_REFUND_REQUEST_SUCCESS:
-		case WOOCOMMERCE_ORDER_REFUND_REQUEST_FAILURE:
-			return Object.assign( {}, state, { [ action.orderId ]: WOOCOMMERCE_ORDER_REFUND_REQUEST === action.type } );
 		default:
 			return state;
 	}
@@ -136,7 +114,9 @@ export function totalPages( state = 1, action ) {
 
 export default combineReducers( {
 	isQueryLoading,
+	isLoading,
 	items,
 	queries,
+	refunds,
 	totalPages,
 } );
