+++ client/extensions/woocommerce/state/sites/orders/refunds/test/actions.js
@@ -1,0 +1,72 @@
+/**
+ * External dependencies
+ */
+import { expect } from 'chai';
+import { spy } from 'sinon';
+
+/**
+ * Internal dependencies
+ */
+import { sendRefund } from '../actions';
+import useNock from 'test/helpers/use-nock';
+import { useSandbox } from 'test/helpers/use-sinon';
+import {
+	WOOCOMMERCE_ORDER_REFUND_CREATE,
+	WOOCOMMERCE_ORDER_REFUND_CREATE_SUCCESS,
+} from 'woocommerce/state/action-types';
+import order from '../../test/fixtures/order';
+
+describe( 'actions', () => {
+	describe( '#sendRefund()', () => {
+		const siteId = '123';
+		const refundObj = {
+			amount: '10',
+			reason: 'Testing reason.',
+		};
+
+		useSandbox();
+		useNock( ( nock ) => {
+			nock( 'https://public-api.wordpress.com:443' )
+				.persist()
+				.post( '/rest/v1.1/jetpack-blogs/123/rest-api/' )
+				.query( { path: '/wc/v3/orders/40/refunds&_method=get', json: true, body: refundObj } )
+				.reply( 200, {
+					data: order,
+				} )
+				.post( '/rest/v1.1/jetpack-blogs/234/rest-api/' )
+				.query( { path: '/wc/v3/orders/invalid/refunds&_method=get', json: true, body: refundObj } )
+				.reply( 404, {
+					data: {
+						message: 'No route was found matching the URL and request method',
+						error: 'rest_no_route',
+						status: 400,
+					}
+				} );
+		} );
+
+		it( 'should dispatch an action', () => {
+			const getState = () => ( {} );
+			const dispatch = spy();
+			sendRefund( siteId, 40, refundObj )( dispatch, getState );
+			expect( dispatch ).to.have.been.calledWith( {
+				type: WOOCOMMERCE_ORDER_REFUND_CREATE,
+				siteId,
+				orderId: 40,
+			} );
+		} );
+
+		it( 'should dispatch a success action with the order when the refund request completes', () => {
+			const getState = () => ( {} );
+			const dispatch = spy();
+			const response = sendRefund( siteId, 40, refundObj )( dispatch, getState );
+
+			return response.then( () => {
+				expect( dispatch ).to.have.been.calledWith( {
+					type: WOOCOMMERCE_ORDER_REFUND_CREATE_SUCCESS,
+					siteId,
+					orderId: 40,
+				} );
+			} );
+		} );
+	} );
+} );
