+++ client/extensions/woocommerce/state/sites/orders/selectors.js
@@ -58,18 +58,6 @@ export const isOrderLoading = ( state, orderId, siteId = getSelectedSiteId( stat
 
 /**
  * @param {Object} state Whole Redux state tree
- * @param {Number} orderId Order ID to check
- * @param {Number} [siteId] Site ID to check. If not provided, the Site ID selected in the UI will be used
- * @return {boolean} Whether an order refund has been requested (or completed)
- */
-export const isOrderRefunding = ( state, orderId, siteId = getSelectedSiteId( state ) ) => {
-	const isRefunding = get( state, [ 'extensions', 'woocommerce', 'sites', siteId, 'orders', 'isRefunding', orderId ] );
-	// Strict check because it could also be undefined.
-	return ( true === isRefunding );
-};
-
-/**
- * @param {Object} state Whole Redux state tree
  * @param {Number} [page] Page of orders. If not provided, defaults to first page.
  * @param {Number} [siteId] Site ID to check. If not provided, the Site ID selected in the UI will be used
  * @return {array|false} List of orders, or false if there was an error
